fix: move non page schema out of the pages folder
